+++ test/templates/index.html
@@ -8,7 +8,7 @@
 <body>
     <div class="container">
         <h1>🚀 Quick Python Web App Prototype</h1>
-        <input type="text" id="nameInput" placeholder="Enter your name">
+        <input type="text" id="nameInput" placeholder="post">
         <button onclick="sendMessage()">Submit</button>
         <p id="response"></p>
     </div>
@@ -27,3 +27,5 @@
     </script>
 </body>
 </html>
+<!DOCTYPE html>
+<html lang="en">
